integration: fix metalint
